fix: remove data file from pages
