--- src/pages/profile/data.tsx
@@ -1,12 +1,0 @@
-export const profileFields = [
-  {
-    id: "name",
-    title: "Name",
-    placeholder: "Enter Name",
-  },
-  {
-    id: "email",
-    title: "Email",
-    placeholder: "Enter email",
-  },
-];
