--- src/pages/project/[id]/details/members/data.tsx
@@ -1,12 +1,0 @@
-export const memberColumns = [
-  {
-    title: "Name",
-    key: "name",
-    render: (data: any) => data.user.name,
-  },
-  {
-    title: "Email",
-    key: "email",
-    render: (data: any) => data.user.email,
-  },
-];
